Move ancestors/descendants out of timelines reducer
